moved auth files around
